some rearranging
