+++ samples/chaincode/spaces/src/container-services/manager.ts
@@ -1,3 +1,8 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
 import { EventEmitter } from "events";
 import { BaseContainerService } from "@microsoft/fluid-aqueduct";
